2023/11/03  Bug 修复：分类、标签删除接口添加是否关联文章校验

+++ weblog-module-admin/src/main/java/com/youngzz1k/weblog/admin/service/impl/AdminCategoryServiceImpl.java
@@ -8,7 +8,9 @@ import com.youngzz1k.weblog.admin.model.vo.category.FindCategoryPageListReqVO;
 import com.youngzz1k.weblog.admin.model.vo.category.FindCategoryPageListRspVO;
 import com.youngzz1k.weblog.admin.model.vo.category.AddCategoryReqVO;
 import com.youngzz1k.weblog.admin.service.AdminCategoryService;
+import com.youngzz1k.weblog.common.domain.dos.ArticleCategoryRelDO;
 import com.youngzz1k.weblog.common.domain.dos.CategoryDO;
+import com.youngzz1k.weblog.common.domain.mapper.ArticleCategoryRelMapper;
 import com.youngzz1k.weblog.common.domain.mapper.CategoryMapper;
 import com.youngzz1k.weblog.common.enums.ResponseCodeEnum;
 import com.youngzz1k.weblog.common.exception.BizException;
@@ -32,6 +34,9 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private ArticleCategoryRelMapper articleCategoryRelMapper;
 
     /**
      * 添加分类
@@ -105,6 +110,14 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     public Response deleteCategory(DeleteCategoryReqVO deleteCategoryReqVO) {
         // 分类 ID
         Long categoryId = deleteCategoryReqVO.getId();
+
+        // 校验该分类下是否已经有文章，若有，则提示需要先删除分类下所有文章，才能删除
+        ArticleCategoryRelDO articleCategoryRelDO = articleCategoryRelMapper.selectOneByCategoryId(categoryId);
+
+        if (Objects.nonNull(articleCategoryRelDO)) {
+            log.warn("==> 此分类下包含文章，无法删除，categoryId: {}", categoryId);
+            throw new BizException(ResponseCodeEnum.CATEGORY_CAN_NOT_DELETE);
+        }
 
         // 删除分类
         categoryMapper.deleteById(categoryId);
